+++ src/index.ts
@@ -6,3 +6,4 @@ export * from "./array";
 export * from "./invariants";
 export * from "./contractDecorators";
 export * from "./utilityTypes/index";
+