+++ src/utilityTypes/index.ts
@@ -2,3 +2,5 @@ export type { WithToString } from "./WithToString";
 export type { ValueObject } from "./ValueObject";
 export { Timestamp } from "./Timestamp";
 export type { Invariant, TestFunction } from "./Invariants";
+export type { Unit } from "./Unit";
+export { unit } from "./Unit";